moved more refactoring
