+++ cvat-ui/src/components/analytics-page/task-quality/empty-job.tsx
