+++ cvat-ui/src/components/analytics-page/task-quality/gt-conflicts.tsx
