+++ cvat-ui/src/components/analytics-page/task-quality/issues.tsx
