+++ cvat-ui/src/components/analytics-page/task-quality/job-list.tsx
