+++ cvat-ui/src/components/analytics-page/task-quality/mean-quality.tsx
