+++ cvat-ui/src/components/analytics-page/task-quality/task-quality-component.tsx
@@ -17,7 +17,7 @@ import GtConflicts from './gt-conflicts';
 import Issues from './issues';
 import JobList from './job-list';
 import MeanQuality from './mean-quality';
-import QualitySettingsModal from './quality-settings-modal';
+import QualitySettingsModal from '../shared/quality-settings-modal';
 
 const core = getCore();
 
@@ -128,7 +128,6 @@ function TaskQualityComponent(props: Props): JSX.Element {
                             setQualitySettings={setQualitySettings}
                             visible={qualitySettingsVisible}
                             setVisible={setQualitySettingsVisible}
-                            redirectToProjectId={task.projectId}
                         />
                     </>
                 )
